11.16.0
Add "OrderedList" and "UnorderedList" classes, subclasses of "list", to
"compare" to allow us to know whether a list is intended to be compared
with order in mind or not, rather than rely on a global type_compare.
This allows some better mixing when we have multiple levels of nesting.
  As I always forget, unordered list comparisons of nested objects can't
give us very useful diffs, so tests should always use unordered
comparisons sparingly.